refactor: 大规模迁移系统调用策略到services层
- 拆分并迁移大量系统调用策略代码到对应services子模块，统一代码组织
- 重构mprotect/madvise_mlock等模块，将实现从framework层迁移到services层并保留兼容导出
- 新增proc/fs/credo/mm等子模块的策略实现文件，实现职责分离
- 替换旧日志宏为slog系列宏，统一日志输出
- 修复madvise/mlock系统调用的参数类型转换问题

+++ src/kernel/services/proc/madvise_mlock.rs
@@ -209,7 +209,7 @@ pub type MlockResult<T> = Result<T, MlockError>;
 ///
 /// 返回 0 = 成功, MlockError 表示失败原因
 pub fn madvise(addr: usize, len: usize, advice: Advice) -> MlockResult<()> {
-    let rc = fw_ml::sys_madvise(addr as u64, len as u64, advice.as_u32());
+    let rc = fw_ml::sys_madvise(addr as u64, len as u64, advice.as_u32() as u64);
     if rc == 0 {
         Ok(())
     } else {
@@ -249,7 +249,7 @@ pub fn munlock(addr: usize, len: usize) -> MlockResult<()> {
 
 /// 进程级锁定 (MCL_CURRENT 锁现有 / MCL_FUTURE 锁未来)
 pub fn mlockall(flags: MlockAllFlags) -> MlockResult<()> {
-    let rc = fw_ml::sys_mlockall(flags.bits());
+    let rc = fw_ml::sys_mlockall(flags.bits() as u64);
     if rc == 0 {
         Ok(())
     } else {
